Allow propper keyboard navigation in the toolbar
This commit updates the toolbar button and the popup to allow a proper
keyboard tab navigation.

The problem with the tab navigation is that the editor uses a shadow
root in the end of the html to put popups, this change put the popup
inside the span that contains each toolbar button if the config
allowTabNavigation is active.